Move store to ApiItem level
