+++ packages/docusaurus-theme-openapi-docs/src/theme/ApiDemoPanel/ContentType/index.tsx
@@ -7,7 +7,7 @@
 
 import React from "react";
 
-import { useTypedDispatch, useTypedSelector } from "../hooks";
+import { useTypedDispatch, useTypedSelector } from "../../ApiItem/hooks";
 import FormItem from "./../FormItem";
 import FormSelect from "./../FormSelect";
 import { setContentType } from "./slice";
