+++ packages/docusaurus-theme-openapi-docs/src/theme/ApiDemoPanel/Server/index.tsx
@@ -8,7 +8,7 @@
 import React, { useState } from "react";
 
 import FloatingButton from "../FloatingButton";
-import { useTypedDispatch, useTypedSelector } from "../hooks";
+import { useTypedDispatch, useTypedSelector } from "../../ApiItem/hooks";
 import FormItem from "./../FormItem";
 import FormSelect from "./../FormSelect";
 import FormTextInput from "./../FormTextInput";
