+++ packages/docusaurus-theme-openapi-docs/src/theme/ApiItem/hooks.ts
